+++ renderer/geometry.py
@@ -7,6 +7,7 @@ import jax.lax as lax
 import jax.numpy as jnp
 from jaxtyping import Array, Float, Integer, Num, jaxtyped
 
+from ._meta_utils import add_tracing_name
 from .types import Triangle2Df, Vec2f, Vec3f, Vec4f
 
 # Transform matrix that takes a batch of homogeneous 3D vertices and transform
@@ -28,7 +29,8 @@ Viewport = Float[Array, "4 4"]
 
 
 @jaxtyped
-@partial(jax.jit, donate_argnums=(0, ))
+@partial(jax.jit, donate_argnums=(0, ), inline=True)
+@add_tracing_name
 def normalise(vector: Float[Array, "dim"]) -> Float[Array, "dim"]:
     """normalise vector in-place."""
     return vector / jnp.linalg.norm(vector)
@@ -48,7 +50,8 @@ class Interpolation(enum.Enum):
     """Perspective correction: linear interpolation in clip space"""
 
     @jaxtyped
-    @partial(jax.jit, static_argnames=("self", ))
+    @partial(jax.jit, static_argnames=("self", ), inline=True)
+    @add_tracing_name
     def __call__(
         self,
         values: Num[Array, "3 *valueDimensions"],
@@ -92,7 +95,8 @@ class Interpolation(enum.Enum):
 
 
 @jaxtyped
-@partial(jax.jit, static_argnames=("mode", ))
+@partial(jax.jit, static_argnames=("mode", ), inline=True)
+@add_tracing_name
 def interpolate(
     values: Num[Array, "3 *valueDimensions"],
     barycentric_screen: Vec3f,
@@ -115,7 +119,8 @@ def interpolate(
 
 
 @jaxtyped
-@jax.jit
+@partial(jax.jit, inline=True)
+@add_tracing_name
 def to_homogeneous(
     coordinates: Float[Array, "*batch dim"],
     value: Float[Array, "*batch"] = jnp.array(1.),
@@ -138,7 +143,8 @@ def to_homogeneous(
 
 
 @jaxtyped
-@jax.jit
+@partial(jax.jit, inline=True)
+@add_tracing_name
 def normalise_homogeneous(
     coordinates: Float[Array, "*batch dim"], ) -> Float[Array, "*batch dim"]:
     """Transform the homogenous coordinates to make the scale factor equals to
@@ -152,7 +158,8 @@ def normalise_homogeneous(
 
 
 @jaxtyped
-@jax.jit
+@partial(jax.jit, inline=True)
+@add_tracing_name
 def to_cartesian(
     coordinates: Float[Array, "*batch dim"], ) -> Float[Array, "*batch dim-1"]:
     """Transform the homogenous coordinates to cartesian coordinates by divide
@@ -192,7 +199,8 @@ class Camera(NamedTuple):
 
     @classmethod
     @jaxtyped
-    @partial(jax.jit, static_argnames=("cls", ))
+    @partial(jax.jit, static_argnames=("cls", ), inline=True)
+    @add_tracing_name
     def create(
         cls,
         view: View,
@@ -240,7 +248,8 @@ class Camera(NamedTuple):
 
     @staticmethod
     @jaxtyped
-    @jax.jit
+    @partial(jax.jit, inline=True)
+    @add_tracing_name
     def apply(
         points: Num[Array, "*N 4"],
         matrix: Num[Array, "4 4"],
@@ -275,7 +284,8 @@ class Camera(NamedTuple):
 
     @classmethod
     @jaxtyped
-    @partial(jax.jit, static_argnames=("cls", ))
+    @partial(jax.jit, static_argnames=("cls", ), inline=True)
+    @add_tracing_name
     def apply_pos(
         cls,
         points: Num[Array, "*N 3"],
@@ -305,7 +315,8 @@ class Camera(NamedTuple):
 
     @classmethod
     @jaxtyped
-    @partial(jax.jit, static_argnames=("cls", ))
+    @partial(jax.jit, static_argnames=("cls", ), inline=True)
+    @add_tracing_name
     def apply_vec(
         cls,
         vectors: Num[Array, "*N 3"],
@@ -343,7 +354,8 @@ class Camera(NamedTuple):
         return transformed_normalised
 
     @jaxtyped
-    @jax.jit
+    @partial(jax.jit, inline=True)
+    @add_tracing_name
     def to_screen(
         self,
         points: Num[Array, "*N 4"],
@@ -368,7 +380,8 @@ class Camera(NamedTuple):
         return normalised
 
     @jaxtyped
-    @jax.jit
+    @partial(jax.jit, inline=True)
+    @add_tracing_name
     def to_clip(
         self,
         points: Num[Array, "*N 4"],
@@ -390,7 +403,8 @@ class Camera(NamedTuple):
         return clip_space
 
     @jaxtyped
-    @jax.jit
+    @partial(jax.jit, inline=True)
+    @add_tracing_name
     def to_screen_inv(
         self,
         screen: Float[Array, "*N 4"],
@@ -442,7 +456,8 @@ class Camera(NamedTuple):
 
     @staticmethod
     @jaxtyped
-    @jax.jit
+    @partial(jax.jit, inline=True)
+    @add_tracing_name
     def inv_scale_translation_matrix(
             scale_translation_mat: Float[Array, "4 4"]) -> Float[Array, "4 4"]:
         """Compute the inverse matrix of a (4, 4) matrix representing a scale and translation, in a form of:
@@ -483,7 +498,8 @@ class Camera(NamedTuple):
 
     @staticmethod
     @jaxtyped
-    @jax.jit
+    @partial(jax.jit, inline=True)
+    @add_tracing_name
     def view_matrix(
         eye: Vec3f,
         centre: Vec3f,
@@ -524,7 +540,8 @@ class Camera(NamedTuple):
 
     @staticmethod
     @jaxtyped
-    @jax.jit
+    @partial(jax.jit, inline=True)
+    @add_tracing_name
     def view_matrix_inv(
         eye: Vec3f,
         centre: Vec3f,
@@ -577,7 +594,8 @@ class Camera(NamedTuple):
 
     @staticmethod
     @jaxtyped
-    @jax.jit
+    @partial(jax.jit, inline=True)
+    @add_tracing_name
     def perspective_projection_matrix(
         fovy: jnp.floating[Any],
         aspect: jnp.floating[Any],
@@ -621,7 +639,8 @@ class Camera(NamedTuple):
 
     @classmethod
     @jaxtyped
-    @partial(jax.jit, static_argnames=("cls", ))
+    @partial(jax.jit, static_argnames=("cls", ), inline=True)
+    @add_tracing_name
     def perspective_projection_matrix_inv(cls, mat: Projection) -> Projection:
         """Create the inverse of a perspective projection matrix as defined in
             `perspective_projection_matrix`.
@@ -650,7 +669,8 @@ class Camera(NamedTuple):
 
     @staticmethod
     @jaxtyped
-    @jax.jit
+    @partial(jax.jit, inline=True)
+    @add_tracing_name
     def orthographic_projection_matrix(
         left: jnp.floating[Any],
         right: jnp.floating[Any],
@@ -694,6 +714,7 @@ class Camera(NamedTuple):
     @classmethod
     @jaxtyped
     @partial(jax.jit, static_argnames=("cls", ))
+    @add_tracing_name
     def orthographic_projection_matrix_inv(cls, mat: Projection) -> Projection:
         """Create the inverse of a orthographic projection matrix as defined in
             `orthographic_projection_matrix`. Since orthographic projection
@@ -707,7 +728,8 @@ class Camera(NamedTuple):
 
     @staticmethod
     @jaxtyped
-    @jax.jit
+    @partial(jax.jit, inline=True)
+    @add_tracing_name
     def perspective_projection_matrix_tinyrenderer(
         eye: Vec3f,
         centre: Vec3f,
@@ -734,7 +756,8 @@ class Camera(NamedTuple):
 
     @staticmethod
     @jaxtyped
-    @jax.jit
+    @partial(jax.jit, inline=True)
+    @add_tracing_name
     def viewport_matrix(
         lowerbound: Num[Array, "2"],
         dimension: Integer[Array, "2"],
@@ -766,7 +789,8 @@ class Camera(NamedTuple):
 
     @classmethod
     @jaxtyped
-    @partial(jax.jit, static_argnames=("cls", ))
+    @partial(jax.jit, static_argnames=("cls", ), inline=True)
+    @add_tracing_name
     def viewport_matrix_inv(cls, viewport: Viewport) -> Viewport:
         """Create the inverse of a viewport matrix as defined in `viewport_matrix`.
 
@@ -782,7 +806,8 @@ class Camera(NamedTuple):
 
     @staticmethod
     @jaxtyped
-    @jax.jit
+    @partial(jax.jit, inline=True)
+    @add_tracing_name
     def world_to_screen_matrix(width: int, height: int) -> World2Screen:
         """Generate the projection matrix to convert model coordinates to
             screen/canvas coordinates.
@@ -804,7 +829,8 @@ class Camera(NamedTuple):
 
 
 @jaxtyped
-@jax.jit
+@partial(jax.jit, inline=True)
+@add_tracing_name
 def compute_normal(triangle_verts: Float[Array, "3 3"]) -> Float[Array, "3"]:
     normal: Float[Array, "3"] = jnp.cross(
         triangle_verts[2, :] - triangle_verts[0, :],
@@ -817,13 +843,15 @@ def compute_normal(triangle_verts: Float[Array, "3 3"]) -> Float[Array, "3"]:
 
 
 @jaxtyped
-@jax.jit
+@partial(jax.jit, inline=True)
+@add_tracing_name
 def compute_normals(batch_verts: Float[Array, "b 3 3"]) -> Float[Array, "b 3"]:
     return jax.vmap(compute_normal)(batch_verts)
 
 
 @jaxtyped
-@jax.jit
+@partial(jax.jit, inline=True)
+@add_tracing_name
 def quaternion(
     rotation_axis: Union[Vec3f, tuple[float, float, float]],
     rotation_angle: Union[Float[Array, ""], float],
@@ -850,7 +878,8 @@ def quaternion(
 
 
 @jaxtyped
-@jax.jit
+@partial(jax.jit, inline=True)
+@add_tracing_name
 def quaternion_mul(quatA: Vec4f, quatB: Vec4f) -> Vec4f:
     """Multiply two quaternion rotations, as to composite them.
 
@@ -877,7 +906,8 @@ def quaternion_mul(quatA: Vec4f, quatB: Vec4f) -> Vec4f:
 
 
 @jaxtyped
-@jax.jit
+@partial(jax.jit, inline=True)
+@add_tracing_name
 def rotation_matrix(
     rotation_axis: Union[Vec3f, tuple[float, float, float]],
     rotation_angle: Union[Float[Array, ""], float],
@@ -908,7 +938,8 @@ def rotation_matrix(
 
 
 @jaxtyped
-@jax.jit
+@partial(jax.jit, inline=True)
+@add_tracing_name
 def transform_matrix_from_rotation(rotation: Vec4f) -> Float[Array, "3 3"]:
     """Generate a transform matrix from a quaternion rotation.
 
@@ -936,7 +967,8 @@ def transform_matrix_from_rotation(rotation: Vec4f) -> Float[Array, "3 3"]:
 
 
 @jaxtyped
-@jax.jit
+@partial(jax.jit, inline=True)
+@add_tracing_name
 def barycentric(pts: Triangle2Df, p: Vec2f) -> Vec3f:
     """Compute the barycentric coordinate of `p`.
         Returns u[-1] < 0 if `p` is outside of the triangle.