+++ renderer/shaders/phong_darboux.py
@@ -1,3 +1,4 @@
+from functools import partial
 from typing import NamedTuple
 
 import jax
@@ -6,6 +7,7 @@ import jax.numpy as jnp
 from jax.tree_util import Partial
 from jaxtyping import Array, Bool, Float, Integer, jaxtyped
 
+from .._meta_utils import add_tracing_name
 from ..geometry import (Camera, Interpolation, interpolate, normalise,
                         to_cartesian, to_homogeneous)
 from ..shader import ID, MixerOutput, PerFragment, PerVertex, Shader
@@ -79,7 +81,8 @@ class PhongTextureDarbouxShader(Shader[PhongTextureDarbouxExtraInput,
 
     @staticmethod
     @jaxtyped
-    @jax.jit
+    @partial(jax.jit, inline=True)
+    @add_tracing_name
     def vertex(
         gl_VertexID: ID,
         gl_InstanceID: ID,
@@ -132,7 +135,8 @@ class PhongTextureDarbouxShader(Shader[PhongTextureDarbouxExtraInput,
 
     @staticmethod
     @jaxtyped
-    @jax.jit
+    @partial(jax.jit, inline=True)
+    @add_tracing_name
     def interpolate(
         values: PhongTextureDarbouxExtraFragmentData,
         barycentric_screen: Vec3f,
@@ -167,7 +171,8 @@ class PhongTextureDarbouxShader(Shader[PhongTextureDarbouxExtraInput,
 
     @staticmethod
     @jaxtyped
-    @jax.jit
+    @partial(jax.jit, inline=True)
+    @add_tracing_name
     def fragment(
         gl_FragCoord: Vec4f,
         gl_FrontFacing: Bool[Array, ""],
@@ -236,7 +241,8 @@ class PhongTextureDarbouxShader(Shader[PhongTextureDarbouxExtraInput,
 
     @staticmethod
     @jaxtyped
-    @jax.jit
+    @partial(jax.jit, inline=True)
+    @add_tracing_name
     def mix(
         gl_FragDepth: Float[Array, "primitives"],
         keeps: Bool[Array, "primitives"],