+++ renderer/shaders/phong_reflection_shadow.py
@@ -1,3 +1,4 @@
+from functools import partial
 from typing import NamedTuple
 
 import jax
@@ -5,6 +6,7 @@ import jax.lax as lax
 import jax.numpy as jnp
 from jaxtyping import Array, Bool, Float, Integer, jaxtyped
 
+from .._meta_utils import add_tracing_name
 from ..geometry import Camera, normalise, normalise_homogeneous, to_homogeneous
 from ..model import MergedModel
 from ..shader import ID, MixerOutput, PerFragment, PerVertex, Shader
@@ -85,7 +87,8 @@ class PhongReflectionShadowTextureShader(
 
     @staticmethod
     @jaxtyped
-    @jax.jit
+    @partial(jax.jit, inline=True)
+    @add_tracing_name
     def vertex(
         gl_VertexID: ID,
         gl_InstanceID: ID,
@@ -124,7 +127,8 @@ class PhongReflectionShadowTextureShader(
 
     @staticmethod
     @jaxtyped
-    @jax.jit
+    @partial(jax.jit, inline=True)
+    @add_tracing_name
     def interpolate(
         values: PhongReflectionShadowTextureExtraFragmentData,
         barycentric_screen: Vec3f,
@@ -144,7 +148,8 @@ class PhongReflectionShadowTextureShader(
 
     @staticmethod
     @jaxtyped
-    @jax.jit
+    @partial(jax.jit, inline=True)
+    @add_tracing_name
     def fragment(
         gl_FragCoord: Vec4f,
         gl_FrontFacing: Bool[Array, ""],
@@ -220,11 +225,7 @@ class PhongReflectionShadowTextureShader(
                 use_default_depth=built_in.use_default_depth,
             ),
             PhongReflectionShadowTextureExtraFragmentData(
-                colour=lax.cond(
-                    (colour >= 0).all(),
-                    lambda: colour,
-                    lambda: jnp.zeros(3),
-                ),
+                colour=colour,
                 uv=varying.uv,
                 normal=varying.normal,
             ),
@@ -232,7 +233,8 @@ class PhongReflectionShadowTextureShader(
 
     @staticmethod
     @jaxtyped
-    @jax.jit
+    @partial(jax.jit, inline=True)
+    @add_tracing_name
     def mix(
         gl_FragDepth: Float[Array, "primitives"],
         keeps: Bool[Array, "primitives"],