+++ renderer/utils.py
@@ -1,3 +1,4 @@
+from functools import partial
 from typing import Sequence, Union
 
 import jax
@@ -5,11 +6,13 @@ import jax.numpy as jnp
 from jax import lax
 from jaxtyping import Array, Integer, Num, Shaped, jaxtyped
 
+from ._meta_utils import add_tracing_name
 from .types import Canvas, Texture, ZBuffer
 
 
 @jaxtyped
-@jax.jit
+@partial(jax.jit, inline=True)
+@add_tracing_name
 def get_value_from_index(
     matrix: Shaped[Array, "width height batch *valueDimensions"],
     index: Integer[Array, "width height"],
@@ -19,7 +22,8 @@ def get_value_from_index(
 
 
 @jaxtyped
-@jax.jit
+@partial(jax.jit, inline=True)
+@add_tracing_name
 def merge_canvases(
     zbuffers: Num[Array, "batch width height"],
     canvases: Shaped[Array, "batch width height channel"],
@@ -47,7 +51,8 @@ def merge_canvases(
 
 
 @jaxtyped
-@jax.jit
+@partial(jax.jit, inline=True)
+@add_tracing_name
 def transpose_for_display(
     matrix: Num[Array, "fst snd *channel"],
     flip_vertical: bool = True,
@@ -69,6 +74,7 @@ def transpose_for_display(
 
 
 @jaxtyped
+@add_tracing_name
 def build_texture_from_PyTinyrenderer(
     texture: Union[Num[Array, "length"], Sequence[float]],
     width: int,